move congruence test into its own file
